crafting inventory works, but needs testing
also updated the recipes

+++ item.h
@@ -113,8 +113,8 @@ public:
 
  bool matches_type(const itype_id &type) const;
  int get_charges_of(const itype_id &type) const;
- double get_functionality_time_modi(const itype_id &func) const;
- double get_damaged_modi() const;
+ float get_functionality_time_modi(const itype_id &func) const;
+ float get_damaged_modi() const;
 
  int weight() const;
